Add pathspec dependency
In order to handle gitignore file exclusion patterns, we would need to
either write a lot of custom code that handles all the edge cases of its
syntax. Instead, let's pull in a 13-year-old dependency called pathspec
that can handle it for us.

Another avenue here would be to vendor the dependency, but I'm doing the
simplest thing in this commit.