Introduce a regex exclusion pattern path
Introduce a second way to handle exclusion patterns in the `_match`
function with regex patterns.

This is a midway commit: tests are passing, and it's at a point where
can choose to either standardize using regex-only in `_match` by using
`fnmatch.translate` on the fnmatch patters, or keep it as-is with the
onus of needing to know what kind of pattern it is on the caller.