Add recommended venv directory to gitignore
Following the steps in CONTRIBUTING.md, I've generated a venv called
`vulture_dev/` that sits in the repo root.

As vulture tests its own code as part of its test suite and we're
introducing a way to read gitignores as a default exclusion pattern
source, we also need to include it to this project's files in order for
tests to pass.