Add gitignore support using pathspec
>[!WARNING]
>N.B. Breaks backwards compatibility.

Implements a way for vulture to read a path or projects' gitignore as a
source of filename exclusion patterns is no `--exclude` flag is
explictly passed or none is found in a TOML config.

Uses the same implementation logic as `black.find_project_root`. It
finds the most closest common parent .gitignore to the supplied `paths`,
and supports relative use. For example:

    $ vulture ../another/repo

Will use `../another/repo/.gitignore` if available.